+++ packages/nodes-base/nodes/Microsoft/Sql/test/utils.test.ts
@@ -1,0 +1,3 @@
+describe('MSSQL tests', () => {
+	it('should perform insert operation', () => {});
+});
